# services/api/openapi.json
{
  "components": {
    "schemas": {
...
      "FeedRequest": {
        "properties": {
          "feed_url": {
            "title": "Feed Url",
            "type": "string"
          },
          "source": {
            "default": "feed",
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "feed_url"
        ],
        "title": "FeedRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestRequest": {
        "properties": {
          "cities": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Cities",
            "type": "array"
          },
          "command": {
            "enum": [
              "seed",
              "rents",
              "osm",
              "bootstrap",
              "opendata"
            ],
            "title": "Command",
            "type": "string"
          },
          "sources": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Sources",
            "type": "array"
          }
        },
        "required": [
          "command"
        ],
        "title": "IngestRequest",
        "type": "object"
      },
      "JobResponse": {
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "queued_at": {
            "format": "date-time",
            "title": "Queued At",
            "type": "string"
          },
          "task": {
            "title": "Task",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "task",
          "queued_at"
        ],
        "title": "JobResponse",
        "type": "object"
      },
...
      "UserCreate": {
        "description": "DTO for user registration.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "password": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "UserCreate",
        "type": "object"
      },
      "UserLogin": {
        "description": "DTO for user login.",
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "UserLogin",
        "type": "object"
      },
      "UserResponse": {
        "description": "DTO for returning user info (never includes password).",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          }
        },
        "required": [
          "email",
          "id",
          "role",
          "created_at",
          "updated_at"
        ],
        "title": "UserResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Canadian residential investment analytics: listings, comparables, cash flow, markets.",
    "title": "NeighborIQ API",
    "version": "0.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/feeds": {
      "post": {
        "description": "Crawl a licensed partner listing feed (canonical JSON/CSV).",
        "operationId": "trigger_feed_api_v1_admin_feeds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeedRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Feed",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/ingest": {
      "post": {
        "description": "Load seed listings, rent benchmarks, OSM POIs or open-data sources.",
        "operationId": "trigger_ingestion_api_v1_admin_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Trigger Ingestion",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/api/v1/auth/login": {
      "post": {
        "description": "Authenticate and set session cookies.",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/auth/signup": {
      "post": {
        "description": "Register a new user and sign them in.",
        "operationId": "signup_api_v1_auth_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Signup",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}